19.08 16:37 poprawiono strone główną oraz panel gracza i panel gier

+++ QuickGuess/DTOs/Game/PlayerSongStatsDto.cs
@@ -7,7 +7,7 @@
         public int Incorrect { get; set; }
         public double WinRate { get; set; }
         public int RankingPosition { get; set; }
-        public int BestTimeSec { get; set; }
+        public double BestTimeSec { get; set; }
         public double AvgTimeSec { get; set; }
         public int TotalScore { get; set; }
     }
